# frontend/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Payknot Event Checkout API",
    "version": "1.1.0",
    "description": "API for event management and Solana USDC checkout sessions. Supports cookie auth for host dashboard and Ed25519 signature auth for external agents."
  },
  "servers": [
    {
      "url": "https://pay.crea8r.xyz/api",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "AgentSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Agent-Signature",
        "description": "Agent signature auth. Also send X-Agent-Id and X-Agent-Timestamp. Canonical string: METHOD + \\n + PATH + \\n + TIMESTAMP + \\n + SHA256_HEX(BODY_RAW)."
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "spw_session",
        "description": "Host dashboard session cookie auth."
      },
      "PersonalAccessToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Host-issued personal access token for agent/API usage."
      }
    },
...
    "schemas": {
...
      "AgentPATCreateRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "description": "Human-readable token label.", "example": "Codex production runtime" },
          "expiresInDays": { "type": "integer", "description": "Optional PAT lifetime in days. 0 means no explicit expiry.", "example": 90 }
        }
      },
      "AgentPATRevokeRequest": {
        "type": "object",
        "required": ["tokenId"],
        "properties": {
          "tokenId": { "type": "string", "format": "uuid", "description": "PAT identifier to revoke." }
        }
      },
      "AgentPATListResponse": {
        "type": "object",
        "properties": {
          "tokens": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "name": { "type": "string" },
                "tokenHint": { "type": "string" },
                "scope": { "type": "string" },
                "createdAt": { "type": "string", "format": "date-time" },
                "lastUsedAt": { "type": "string", "format": "date-time" },
                "expiresAt": { "type": "string", "format": "date-time" },
                "revokedAt": { "type": "string", "format": "date-time" }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/agent/pats": {
      "get": {
        "operationId": "listAgentPATs",
        "summary": "List host personal access tokens",
        "description": "Returns host-created PAT metadata. Host session auth required.",
        "security": [{ "SessionCookie": [] }],
        "responses": {
          "200": {
            "description": "PAT list.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AgentPATListResponse" } } }
          },
          "401": { "description": "Unauthorized." }
        }
      },
      "post": {
        "operationId": "createAgentPAT",
        "summary": "Create host personal access token",
        "description": "Creates a new PAT for agent/API usage and returns the token value exactly once. Host session auth required.",
        "security": [{ "SessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AgentPATCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PAT created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "tokenId": { "type": "string", "format": "uuid" },
                    "name": { "type": "string" },
                    "token": { "type": "string" },
                    "scope": { "type": "string" },
                    "expiresAt": { "type": "string", "format": "date-time" },
                    "createdAt": { "type": "string", "format": "date-time" },
                    "tokenHint": { "type": "string" },
                    "ownerEmail": { "type": "string", "format": "email" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation failed." },
          "401": { "description": "Unauthorized." }
        }
      }
    },
    "/agent/pats/revoke": {
      "post": {
        "operationId": "revokeAgentPAT",
        "summary": "Revoke host personal access token",
        "description": "Revokes an existing PAT. Host session auth required.",
        "security": [{ "SessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AgentPATRevokeRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PAT revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "tokenId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "400": { "description": "Missing token id." },
          "401": { "description": "Unauthorized." }
        }
      }
    },
...
  }
}